requirements.txt
Django==4.2.24
Pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
